Added a new method for writing to external OutputStream
Often you may want to externalize the Buffer to something else. However,
we don't want to clone the underlying byte-array in order to do so and
as such, this method will take an OutpuStream as an argument and write
to it.

Note: to be fair, writing to the OutputStream actually gives out the
internal reference to the internal byte-array so a malicious
OutputStream could actually mess witn our internal structure. Still not
sure I want to make a safe copy first to 100% ensure our immutability...