Separating Readable and Writable Buffers into two separate interfaces.
Depening on your needs, it may be that you only need to create a new
buffer and write it with your content but not also read from it. This is
common if you are building network based applications where you build up
the buffer you are about to write to the socket and as such, you do not
need the ability to also read from that buffer. As such, the
WritableBuffer no longer inherits from the ReadableBuffer but there is a
new ReadWriteBuffer that extends both of them.

By doing this, the WritableBuffer is essentially a "builder" for the
underlying byte-array and when you are done, you "lock" it in by
building it, which at that point the WritableBuffer is no longer usabe
and as such, we are guaranteeing that the data, once constructed, cannot
be changed. A very important aspect for most programs. I.e.,
Immutability for the win!